requirements.txt
django==4.2.19
django-ckeditor-5==0.2.17
django-environ==0.12.0
django-cors-headers==4.7.0
django-phonenumber-field==8.0.0
pillow==11.1.0
django-storages==1.14.4
django-environ==0.12.0 
djangorestframework==3.15.2
psycopg2==2.9.10
unidecode
